requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0

# Data validation
pydantic[email]==2.10.3
pydantic-settings==2.6.1
email-validator==2.1.0

# OpenAI API
openai==1.55.3

# PDF processing
pymupdf==1.24.14

# DOCX processing (bonus)
python-docx==1.1.2

# Environment variables
python-dotenv==1.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# Testing (bonus)
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
